chat room

+++ src/main/java/net/uglevodov/restapi/controllers/PostsController.java
@@ -5,10 +5,7 @@ import net.uglevodov.restapi.dto.CommentDto;
 import net.uglevodov.restapi.dto.PostDto;
 import net.uglevodov.restapi.entities.*;
 import net.uglevodov.restapi.security.UserPrincipal;
-import net.uglevodov.restapi.service.DishesService;
-import net.uglevodov.restapi.service.ImageService;
-import net.uglevodov.restapi.service.PostsService;
-import net.uglevodov.restapi.service.WallsService;
+import net.uglevodov.restapi.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -37,6 +34,9 @@ public class PostsController {
 
     @Autowired
     WallsService wallsService;
+
+    @Autowired
+    ChatRoomService chatRoomService;
 
 
     @GetMapping(value = "/get")
@@ -86,9 +86,14 @@ public class PostsController {
     @DeleteMapping(value = "/delete")
     public ResponseEntity<?> delete(@RequestParam(value = "id") Long id,
                                     @RequestParam(value = "wallId") Long wallId,
+                                    @RequestParam(value = "chatroom") boolean chatRoom,
                                     @AuthenticationPrincipal UserPrincipal principal) {
-        wallsService.removePost(wallsService.get(wallId),postsService.get(id));
-        postsService.delete(id, principal.getId());
+
+        if (chatRoom) chatRoomService.removePost(postsService.get(id));
+        else {
+            wallsService.removePost(wallsService.get(wallId), postsService.get(id));
+            postsService.delete(id, principal.getId());
+        }
 
         return new ResponseEntity<>(new ApiResponse(true, "post deleted, id "+id), HttpStatus.OK);
     }
@@ -132,9 +137,12 @@ public class PostsController {
 
         post = postsService.save(post, principal.getId());
 
-        Wall wall = wallsService.get(postDto.getWallId());
-        wallsService.addPost(wall, post);
-
+        if (postDto.isChatRoomPost())
+            chatRoomService.addPost(post);
+        else {
+            Wall wall = wallsService.get(postDto.getWallId());
+            wallsService.addPost(wall, post);
+        }
         return new ResponseEntity<>(post, HttpStatus.ACCEPTED);
     }
 
@@ -164,9 +172,12 @@ public class PostsController {
 
         post = postsService.save(post, principal.getId());
 
-        Wall wall = wallsService.get(postDto.getWallId());
-        wallsService.updatePost(wall, post);
-
+        if (postDto.isChatRoomPost())
+            chatRoomService.updatePost(post);
+        else {
+            Wall wall = wallsService.get(postDto.getWallId());
+            wallsService.updatePost(wall, post);
+        }
         return new ResponseEntity<>(post, HttpStatus.ACCEPTED);
     }
 }
